feat: split out models into subpackage
