+++ src/dataops/models/censusapiendpoint.py
@@ -2,6 +2,11 @@ import os
 import sys
 import requests
 import polars as pl
+from typing import List, Optional, Annotated
+from urllib.parse import urlparse, parse_qs
+from dotenv import load_dotenv
+from datetime import datetime
+
 from pydantic import (
     BaseModel,
     Field,
@@ -11,33 +16,11 @@ from pydantic import (
     computed_field,
     ValidationError,
 )
-from pydantic_settings import BaseSettings, SettingsConfigDict
-from typing import List, Optional, Annotated
-from urllib.parse import urlparse, parse_qs
-from dotenv import load_dotenv
-from datetime import datetime
-# import re
 
 load_dotenv()
 
 # TODO: convert to pydantic settings and implement
 CENSUS_API_KEY = os.getenv("CENSUS_API_KEY")
-
-
-class ApplicationSettings(BaseSettings):
-    """
-    Defines application settings for interacting with the portal platform and Census API.
-    """
-
-    model_config = SettingsConfigDict(env_file=".env", env_file_encoding="utf-8")
-
-    census_api_key: str = Field("", env="CENSUS_API_KEY")
-    domain: str = Field("", env="DOMAIN")
-    source_id: str = Field("", env="SOURCE_ID")
-    target_id: str = Field("", env="TARGET_ID")
-    socrata_user: str = Field("", env="SOCRATA_USER")
-    socrata_pass: str = Field("", env="SOCRATA_PASS")
-    socrata_token: str = Field("", env="SOCRATA_TOKEN")
 
 
 class CensusAPIEndpoint(BaseModel):
